Add files via upload
Java File Handling using File Writer - Output Stream
Java code reads input from the user and writes specific characters (based on the value of a (i.e.)choice) to a file named “input.txt”. If a is “caps”, it writes uppercase letters; if it’s “small”, it writes lowercase letters; if it’s “number”, it writes digits; and if it’s “special”, it writes special characters. The output is saved in the specified file.